Move vm_registry back to interpreter package and resolve circular dependencies
